editing on MedicineController (index,show,store,destroy) adding middleware for medicine endpoints

+++ app/Http/Controllers/MedicineController.php
@@ -6,8 +6,10 @@ use App\Models\Category;
 use App\Models\Medicine;
 use App\Http\Requests\StoreMedicineRequest;
 use App\Http\Requests\UpdateMedicineRequest;
+use App\Models\StatusMedicine;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
+use Ramsey\Uuid\Type\Integer;
 
 class MedicineController extends Controller
 
@@ -19,15 +21,14 @@ class MedicineController extends Controller
      */
     public function index()
     {
-        return Medicine::all();
-    }
+        //showing all medicines for pharmacist
+        if(!auth()->user()->role)
+            return $this->apiResponse(Medicine::all(), 'medicine fetched successfully');
 
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        //
+        //showing all medicines for admin with status of each Medicine
+        else
+            return $this->apiResponse(Medicine::with('statusMedicines')
+            ->get(), 'medicine fetched successfully');
     }
 
     /**
@@ -35,78 +36,91 @@ class MedicineController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'commercial_name' => 'required',
-            'scientific_name' => 'required',
-            'manufacture_company' => 'required',
-            'price' => 'required']);
+        //access just for admin
+        if(!auth()->user()->role)
+            return $this->apiResponse(null, 'access only for admin');
+
+        $validatedData = $request->validate([
+            'commercial_name' => 'required|string|unique:medicines,commercial_name',
+            'scientific_name' => 'required|string',
+            'manufacture_company' => 'required|string',
+            'category_id' => 'required|integer|exists:categories,id',
+            'price' => 'required|integer',
+            'quantity' => 'required|integer',
+            'expiration_date' => 'required|date_format:Y-m-d'
+        ]);
 
 
         $medicine = Medicine::create([
-            'commercial_name' => $request->input('commercial_name'),
-            'scientific_name' => $request->input('scientific_name'),
-            'manufacture_company' => $request->input('manufacture_company'),
-            'price' => $request->input('price'),
-            'category_id' => $request->input('category_id')
+            'commercial_name' => $validatedData['commercial_name'],
+            'scientific_name' => $validatedData['scientific_name'],
+            'manufacture_company' => $validatedData['manufacture_company'],
+            'quantity' => $validatedData['quantity'],
+            'price' => $validatedData['price'],
+            'category_id' => $validatedData['category_id']
         ]);
 
-        if ($medicine) {
-            return $this->apiResponse($medicine, 'the medicine inserted', 201);
-        }
+        StatusMedicine::create([
+            'medicine_id' => $medicine->id,
+            'quantity' => $validatedData['quantity'],
+            'expiration_date' => $validatedData['expiration_date']
+        ]);
 
-        return $this->apiResponse(null, 'the medicine didn\'t created', 401);
+        return $this->apiResponse($medicine,'medicine stored successfully');
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(string $id)
+    public function show($id)
     {
-        return Medicine::find($id);
-
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(Medicine $medicine)
-    {
-        //
+        //show for pharmacist
+        if(!auth()->user()->role)
+            return $this->apiResponse(Medicine::find($id), 'medicine fetched successfully');
+        //show for admin
+        else
+            return $this->apiResponse(Medicine::with('statusMedicines')
+                ->find($id),'medicine fetched successfully');
     }
 
     /**
      * Update the specified resource in storage.
      */
 
-    public function update(Request $request, $medicine_id)
+    public function update(Request $request, $id)
     {
+        //access just for admin
+        if(!auth()->user()->role)
+            return $this->apiResponse(null, 'access only for admin');
+
+        $medicine = Medicine::find($id);
+
+        if (!$medicine)
+            return $this->apiResponse(null, 'No medicine found with the specified ID');
+
         $request->validate([
-            'scientific_name' => 'required|string|max:25',
-            'category_id' => 'required',
-            'trade_name' => 'required|string|max:25',
-            'company' => 'required|string|max:25',
-            'price' => 'required|max:25'
+            //لانو وانت عم تعدل اسم الدوا اذا حطيت اسم موجود يضربلك ايرور
+            // واذا نفس اسم الدوا يلي عم تعدل عليه مايحاكيك شي "ترجمها انت بكرا بالله :)"
+            'commercial_name' => 'string|unique:medicines,commercial_name,' . $medicine->id,
+            'scientific_name' => 'string',
+            'manufacture_company' => 'string',
+            'category_id' => 'integer|exists:categories,id',
+            'price' => 'integer'
         ]);
 
-        $medicine = Medicine::find($medicine_id);
-        if (!$medicine) {
-
-            return $this->apiResponse($medicine, 'the post not found', 404);
-        }
-
-        $medicine = Medicine::update([
-            'commercial_name' => $request->input('commercial_name'),
-            'scientific_name' => $request->input('scientific_name'),
-            'manufacture_company' => $request->input('manufacture_company'),
-            'price' => $request->input('price'),
-            'category_id' => $request->input('category_id')
+        $medicine->update([
+            //can use but, it's a security mistake
+//            $request->all()
+            //so use these approach instead
+            'commercial_name' => $request->input('commercial_name') ?? $medicine->commercial_name,
+            'scientific_name' => $request->input('scientific_name') ?? $medicine->scientific_name,
+            'manufacture_company' => $request->input('manufacture_company') ?? $medicine->manufacture_company,
+            'price' => $request->input('price') ?? $medicine->price,
+            'category_id' => $request->input('category_id') ?? $medicine->category_id
         ]) ;
 
-        if ($medicine)
-        {
 
-            return $this->apiResponse($medicine, 'the post updated', 201);
-        }
+            return $this->apiResponse($medicine, 'the medicine updated');
 
     }
 
@@ -116,24 +130,35 @@ class MedicineController extends Controller
 
     public function destroy($id)
     {
+        //access just for admin
+        if(!auth()->user()->role)
+            return $this->apiResponse(null, 'access only for admin');
+
         $medicine = Medicine::find($id);
+
         if (!$medicine) {
-            return $this->apiResponse($medicine, 'the medicine not found', 404);
+            return $this->apiResponse($medicine, 'the medicine not found');
         }
+
         $medicine->delete($id);
+
         if ($medicine) {
-            return $this->apiResponse(null, 'the medicine deleted', 200);
+            return $this->apiResponse(null, 'the medicine deleted');
         }
+
     }
 
-    public function search($name)
+    /**
+     * search for medicine by name
+     */
 
+    public function search($name)
     {
-        $medicine = Medicine::Where('commercial_name', 'like', '%' . $name . '%')
+        $medicine = Medicine::where('commercial_name', 'like', '%' . $name . '%')
             ->orwhere('scientific_name', 'like', '%' . $name . '%')
             ->get();
         if ($medicine) {
-            return $this->apiResponse($medicine, 'here ur search', 200);
+            return $this->apiResponse($medicine, 'here ur search');
         }
 
     }
